Add additional wait to make test stable
Needed to wait for the spinner to disappear
Otherwise the test throws stale element exception
Because as you click + or - to add or remove products to the basket, the whole basket element is refreshed
The element is removed and readded and that is why a stale element exception is thrown.
The button element is found before the basket is removed and then when the test tries to click the button element, it is no longer there.
There is a same element created at that point, but from seleniums perspective it is no longer the same element